api/requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
aiounittest==1.4.0
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
contextvars==2.4
flake8==3.8.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.8.5
httptools==0.1.1
httpx==0.11.1
hyperframe==5.2.0
idna==2.10
idna-ssl==1.1.0
immutables==0.14
importlib-metadata==1.7.0
mccabe==0.6.1
more-itertools==8.4.0
multidict==4.7.6
packaging==20.4
paho-mqtt==1.5.0
peewee==3.13.3
peewee-migrations==0.3.19
peewee-seed==0.1.8
peewee-validates==1.0.7
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-sanic==1.6.1
python-dateutil==2.8.1
python-dotenv==0.14.0
PyYAML==5.3.1
requests==2.24.0
rfc3986==1.4.0
sanic==20.6.3
Sanic-Cors==0.10.0.post3
Sanic-Plugins-Framework==0.9.3
secure==0.2.1
six==1.15.0
sniffio==1.1.0
typing-extensions==3.7.4.2
ujson==3.1.0
urllib3==1.25.10
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
wrapt==1.12.1
yarl==1.5.1
zipp==3.1.0
